fix(FILTER): :bug: Fix precision truncated
When the sum of array is 6, the mean is 0.6.
But the res is 0, this will cause the wrong result.

+++ HARDWARE/UTIL/filter.c
@@ -44,7 +44,7 @@ s16 UTIL_s16MoveAverageFilter(FILTER_TypeDef *filter, s16 value, u8 filterSize)
     {
         /* keep the filter->u8Len equal to filterSize */
     }
-    res = filter->s32Sum / filter->u8Len;
+    res = (f32)filter->s32Sum / filter->u8Len;
 
     /* Rounding */
     if (res > 0.0)
